backend/requirements/base.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# CORS
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# Security & Encryption
cryptography==41.0.7

# Environment variables
python-decouple==3.8

# Rate limiting
django-ratelimit==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
